Split image controller out to Image and Images controller
